+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+- Error when updating pages with People properties.
+
 ## [v1.14.0] 2024-04-22
 
 ## Added